+++ packages/veui-theme-x/components/AlertBox.js
@@ -1,12 +1,12 @@
-import '../icons/check-circle-o'
-import '../icons/info-circle-o'
-import '../icons/cross-circle-o'
+import '../icons/check-large-circle-o'
+import '../icons/info-large-circle-o'
+import '../icons/cross-large-circle-o'
 import config from 'veui/managers/config'
 
 config.defaults({
   'alertbox.icons': {
-    success: 'check-circle-o',
-    info: 'info-circle-o',
-    error: 'cross-circle-o'
+    success: 'check-large-circle-o',
+    info: 'info-large-circle-o',
+    error: 'cross-large-circle-o'
   }
 })