+++ packages/veui-theme-x/components/Checkbox.js
@@ -1,10 +1,10 @@
 import '../icons/minus-thick'
-import '../icons/check-thick'
+import '../icons/check'
 import config from 'veui/managers/config'
 
 config.defaults({
   'checkbox.icons': {
     indeterminate: 'minus-thick',
-    checked: 'check-thick'
+    checked: 'check'
   }
 })