+++ packages/veui-theme-x/icons/angle-left.js
@@ -1,11 +1,12 @@
 import Icon from 'veui/components/Icon'
 Icon.register({
   "angle-left": {
-    "width": 96,
-    "height": 96,
+    "width": "96",
+    "height": "96",
     "paths": [
       {
-        "d": "M64.24 67.76a6 6 0 0 1-8.48 8.48l-24-24a6 6 0 0 1 0-8.48l24-24a6 6 0 0 1 8.48 8.48L44.5 48l19.75 19.76z"
+        "d": "M64.24 67.76a6 6 0 0 1-8.48 8.48l-24-24a6 6 0 0 1 0-8.48l24-24a6 6 0 0 1 8.48 8.48L44.5 48l19.75 19.76z",
+        "fillRule": "evenodd"
       }
     ]
   }