+++ packages/veui-theme-x/icons/angle-up.js
@@ -1,11 +1,12 @@
 import Icon from 'veui/components/Icon'
 Icon.register({
   "angle-up": {
-    "width": 96,
-    "height": 96,
+    "width": "96",
+    "height": "96",
     "paths": [
       {
-        "d": "M67.76 64.24a6 6 0 0 0 8.48-8.48l-24-24a6 6 0 0 0-8.48 0l-24 24a6 6 0 0 0 8.48 8.48L48 44.5l19.76 19.75z"
+        "d": "M67.76 64.24a6 6 0 0 0 8.48-8.48l-24-24a6 6 0 0 0-8.48 0l-24 24a6 6 0 0 0 8.48 8.48L48 44.5l19.76 19.75z",
+        "fillRule": "evenodd"
       }
     ]
   }