+++ packages/veui-theme-x/icons/check.js
@@ -1,11 +1,12 @@
 import Icon from 'veui/components/Icon'
 Icon.register({
   "check": {
-    "width": 96,
-    "height": 96,
+    "width": "10",
+    "height": "8",
     "paths": [
       {
-        "d": "M16.64 51.36a6.3 6.3 0 0 0-9.28 8.53L29.3 83.73a6.3 6.3 0 0 0 9.31-.04l50.08-55.47a6.3 6.3 0 0 0-9.36-8.44L33.88 70.1 16.64 51.36z"
+        "d": "M3.44 8L0 4.09l1.01-1.04L3.44 5.9 8.94 0 10 1.07z",
+        "fillRule": "evenodd"
       }
     ]
   }