+++ packages/veui-theme-x/icons/cross-large.js
@@ -1,0 +1,13 @@
+import Icon from 'veui/components/Icon'
+Icon.register({
+  "cross-large": {
+    "width": "14",
+    "height": "14",
+    "paths": [
+      {
+        "d": "M7.78 7L14 .78 13.22 0 7 6.22.78 0 0 .78 6.22 7 0 13.22l.78.78L7 7.78 13.22 14l.78-.78L7.78 7z",
+        "fillRule": "evenodd"
+      }
+    ]
+  }
+})