+++ packages/veui-theme-x/icons/exclamation-large-circle-o.js
@@ -1,0 +1,12 @@
+import Icon from 'veui/components/Icon'
+Icon.register({
+  "exclamation-large-circle-o": {
+    "width": "30",
+    "height": "30",
+    "paths": [
+      {
+        "d": "M15 1a14 14 0 1 1 0 28 14 14 0 0 1 0-28zm0-1a15 15 0 1 0 0 30 15 15 0 0 0 0-30zm-1 17h2V9h-2v8zm0 4h2v-2h-2v2z"
+      }
+    ]
+  }
+})