+++ packages/veui-theme-x/icons/info-circle-o.js
@@ -1,11 +1,11 @@
 import Icon from 'veui/components/Icon'
 Icon.register({
   "info-circle-o": {
-    "width": 30,
-    "height": 30,
+    "width": "16",
+    "height": "16",
     "paths": [
       {
-        "d": "M15 29a14 14 0 1 0 0-28 14 14 0 0 0 0 28zm0 1a15 15 0 1 1 0-30 15 15 0 0 1 0 30zm-1-17v8h2v-8h-2zm0-4v2h2V9h-2z"
+        "d": "M8 1a7 7 0 1 0 0 14A7 7 0 0 0 8 1zm0-1a8 8 0 1 1 0 16A8 8 0 0 1 8 0zM7 6.87v5h2v-5H7zm1-1a1 1 0 1 0 0-2 1 1 0 0 0 0 2z"
       }
     ]
   }