+++ packages/veui-theme-x/icons/minus.js
@@ -1,11 +1,12 @@
 import Icon from 'veui/components/Icon'
 Icon.register({
   "minus": {
-    "width": 96,
-    "height": 96,
+    "width": "96",
+    "height": "96",
     "paths": [
       {
-        "d": "M12 42h72a6 6 0 1 1 0 12H12a6 6 0 1 1 0-12z"
+        "d": "M12 42h72a6 6 0 1 1 0 12H12a6 6 0 1 1 0-12z",
+        "fillRule": "evenodd"
       }
     ]
   }