+++ packages/veui-theme-x/icons/sort-desc.js
@@ -1,11 +1,11 @@
 import Icon from 'veui/components/Icon'
 Icon.register({
   "sort-desc": {
-    "width": 96,
-    "height": 96,
+    "width": "12",
+    "height": "12",
     "paths": [
       {
-        "d": "M25.77 70.24A6 6 0 0 1 30 60h36a6 6 0 0 1 4.25 10.24l-18 18.03a6 6 0 0 1-8.5 0l-18-18.03z"
+        "d": "M0 8.03l.65-.73 2.34 3.17V0h1v12h-1L0 8.03z"
       }
     ]
   }