+++ packages/veui-theme-x/icons/triangle-up.js
@@ -1,11 +1,12 @@
 import Icon from 'veui/components/Icon'
 Icon.register({
   "triangle-up": {
-    "width": 96,
-    "height": 96,
+    "width": "96",
+    "height": "96",
     "paths": [
       {
-        "d": "M70.26 52.78c3.77 3.78 2.06 10.24-3.28 10.24h-38c-5.35 0-7-6.46-3.22-10.24l18-18.03a6 6 0 0 1 8.5 0l18 18.03z"
+        "d": "M70.26 52.78c3.77 3.78 2.06 10.24-3.28 10.24h-38c-5.35 0-7-6.46-3.22-10.24l18-18.03a6 6 0 0 1 8.5 0l18 18.03z",
+        "fillRule": "evenodd"
       }
     ]
   }