+++ packages/veui/src/components/ConfirmBox.vue
@@ -3,7 +3,7 @@
   :priority="priority"
   :closable="false"
   ui="reverse"
-  :overlay-class="realOverlayClass">
+  :overlay-class="mergeOverlayClass('veui-confirm-box')">
   <template slot="title">
     <slot name="title">{{ title }}</slot>
   </template>
@@ -20,7 +20,7 @@ import { pick } from 'lodash'
 import Dialog from './Dialog'
 import Button from './Button'
 import config from '../managers/config'
-import { normalizeCSSClass } from '../utils/helper'
+import { overlay } from '../mixins'
 
 config.defaults({
   'confirmbox.priority': 100
@@ -32,8 +32,9 @@ export default {
     'veui-dialog': Dialog,
     'veui-button': Button
   },
+  mixins: [overlay],
   props: {
-    ...pick(Dialog.props, ['open', 'title', 'overlayClass']),
+    ...pick(Dialog.props, ['open', 'title']),
     ui: {
       type: String,
       default: 'reverse'
@@ -44,14 +45,6 @@ export default {
       localOpen: this.open,
       localTitle: this.title,
       priority: config.get('confirmbox.priority')
-    }
-  },
-  computed: {
-    realOverlayClass () {
-      return {
-        ...normalizeCSSClass(this.overlayClass),
-        'veui-confirm-box': true
-      }
     }
   },
   watch: {