+++ packages/veui/src/components/PromptBox.vue
@@ -1,5 +1,5 @@
 <template>
-<veui-dialog :overlay-class="realOverlayClass"
+<veui-dialog :overlay-class="mergeOverlayClass('veui-prompt-box')"
   :ui="ui"
   :open.sync="localOpen"
   :priority="priority"
@@ -18,7 +18,7 @@ import Input from './Input'
 import Dialog from './Dialog'
 import { pick, extend } from 'lodash'
 import config from '../managers/config'
-import { normalizeCSSClass } from '../utils/helper'
+import { overlay } from '../mixins'
 
 config.defaults({
   'promptbox.priority': 100
@@ -30,8 +30,9 @@ export default {
     'veui-input': Input,
     'veui-dialog': Dialog
   },
+  mixins: [overlay],
   props: extend(
-    pick(Dialog.props, ['open', 'title', 'ui', 'overlayClass']),
+    pick(Dialog.props, ['open', 'title', 'ui']),
     {
       content: {
         type: String,
@@ -48,14 +49,6 @@ export default {
       localOpen: this.open,
       priority: config.get('promptbox.priority'),
       localValue: this.value
-    }
-  },
-  computed: {
-    realOverlayClass () {
-      return {
-        ...normalizeCSSClass(this.overlayClass),
-        'veui-prompt-box': true
-      }
     }
   },
   watch: {