+++ packages/veui/src/components/Select/Select.vue
@@ -19,7 +19,8 @@
     v-if="expanded"
     target="button"
     :open="expanded"
-    :options="overlay">
+    :options="overlay"
+    :overlay-class="overlayClass">
     <div ref="box" class="veui-select-options" :ui="ui" v-outside:button="close">
       <slot>
         <template v-for="(option, index) in realOptions">
@@ -60,11 +61,11 @@ import Icon from '../Icon'
 import Button from '../Button'
 import Option from './Option'
 import Overlay from '../Overlay'
-import { input, dropdown, icons } from '../../mixins'
+import { input, dropdown, icons, overlay } from '../../mixins'
 
 export default {
   name: 'veui-select',
-  mixins: [input, dropdown, icons],
+  mixins: [input, dropdown, icons, overlay],
   model: {
     event: 'change'
   },